src/services/billing-microservice/internal/transport/grpc/server_test.go
package grpc

import (
	"context"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
	"go.uber.org/zap"
	"testing"
)

type MockLogger struct {
	mock.Mock
}

func (m *MockLogger) Info(ctx context.Context, msg string, fields ...zap.Field) {
	m.Called(ctx, msg, fields)
}

func (m *MockLogger) Error(ctx context.Context, msg string, fields ...zap.Field) {
	m.Called(ctx, msg, fields)
}

func (m *MockLogger) Sync() error {
	args := m.Called()
	return args.Error(0)
}

func TestNewServer_Failure(t *testing.T) {
	// Test with nil service
	ctx := context.Background()
	server, err := New(ctx, 50051, 8080, nil)
	assert.Error(t, err)
	assert.Nil(t, server)
}
